# api/openapi-spec/openapi.yaml
openapi: 3.0.4
info:
  title: Room Service
  description: |-
    This backend microservice manages 
    - groups formed by registered attendees
    - rooms
    - room assignments of groups or individual attendees
    - a countdown to reveal a secret for external room booking
    
    USE CASE 1: youth hostel style convention
    
    Room groups as provided by this service are useful if you run your convention in a youth hostel, where the registration
    fee includes a bed in one of the rooms. Because each bed costs the same and is included in the convention fee, 
    Attendees can form groups of their own choice, and a few days before the convention, an admin makes the room assignments.
    
    In this model, attendees form groups after they are approved. Each ticket includes a bed anyway, so the only thing
    being managed is who goes in which room.

    USE CASE 2: hotel room lottery
    
    Rooms can also be used in standalone fashion, only assigning individual attendees. This is
    particularly useful if you run your convention in a hotel, and you want to run a room lottery, where
    you are re-selling a fixed set of rooms of different types.
    
    In this model, the cost of accommodation is not included 
    with the convention fee, but instead the room types are separate packages. In order to conduct the
    lottery, registrations that have selected the same room package need to refer to each other (which is an implicit
...
    to the room.

...

...
    
...

...
    
...
servers:
...
paths:
  /groups:
    get:
      tags:
        - groups
      summary: list groups
      description: |-
        Obtain a list of all groups, including their members.       
        
        Admin or Api Key authorization only.
      operationId: listGroups
      parameters:
        - name: member_ids
          in: query
          description: a comma separated list of badge numbers. The result will be limited to all groups that contain at least one of these badge numbers as members.
          schema:
            type: string
            example: 4,11,2560
        - name: min_size
          in: query
          description: list only groups that currently have at least this many members (optional, defaults to 0 which means no limitation)
          schema:
            type: integer
            example: 2
        - name: max_size
          in: query
          description: list only groups that currently have at most this many members (optional, defaults to 0 which means no limitation)
          schema:
            type: integer
            example: 4
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupList'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to see the list of groups (admin/api key only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    post:
      tags:
        - groups
      summary: create a new group
      description: |-
        Create a new group, setting yourself as the owner.
        
        This also adds you as the first member of the group.
        
        You must have a valid registration.
        
        Note that the members and invites fields are ignored. The group is always created with just you as the owner
        (or for admins, if a different owner is specified via badge number, that owner).
      operationId: createGroup
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreate'
        required: true
      responses:
        '201':
          description: Successfully created
          headers:
            Location:
              schema:
                type: string
              description: URL of the created resource, ending in the assigned uuid.
        '400':
          description: Invalid input (name too long/missing, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to create a group (maybe not registered?)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: duplicate (same group name, or you already are in a group)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
  /groups/my:
    get:
      tags:
        - groups
      summary: find my group
      description: |-
        Obtain the group you are in. Must have a valid registration.
        
        This works even for admins, still gives them the group they are in.
        
        Because the user identity is taken from the logged in user, this does not work for Api Key authorization.
        Use the /groups endpoint with member_id parameter instead.
      operationId: findMyGroup
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to see this (maybe not a valid registration?)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: You are not in any groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
  /groups/{uuid}:
    get:
      tags:
        - groups
      summary: obtain a group by uuid
      description: Returns a single group
      operationId: getGroupById
      parameters:
        - name: uuid
          in: path
          description: uuid of the group to return
          required: true
          schema:
            type: string
            example: 604f5ea8-d146-4aac-9a15-4dc33a84eb59
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to see this group (e.g. not a member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    put:
      tags:
        - groups
      summary: update an existing group
      description: |-
        Update an existing group by uuid. Note that you cannot use this to change the group members!
        
        Admins or the current group owner can change the group owner to any member of the group.
      operationId: updateGroup
      parameters:
        - name: uuid
          in: path
          description: uuid of the group to update
          required: true
          schema:
            type: string
            example: 604f5ea8-d146-4aac-9a15-4dc33a84eb59
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
        required: true
      responses:
        '200':
          description: Successful operation
          headers:
            Location:
              schema:
                type: string
              description: URL of the resource, ending in the assigned uuid.
        '400':
          description: Invalid uuid supplied or invalid data in request body, or new owner not part of the group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to change this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: your changes would turn this group into a duplicate (same name as existing other group)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    delete:
      tags:
        - groups
      summary: disband an existing group
      description: |-
        Disband (and delete) an existing group by uuid. Note that this will first kick everyone from the group!
        
        Admins or the current group owner can do this.
      operationId: deleteGroup
      parameters:
        - name: uuid
          in: path
          description: uuid of the group to disband
          required: true
          schema:
            type: string
            example: 604f5ea8-d146-4aac-9a15-4dc33a84eb59
      responses:
        '204':
          description: Successful operation
        '400':
          description: Invalid uuid supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Authorization required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: You do not have permission to do this (not owner/admin).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: An unexpected error occurred. This includes database errors. A best effort attempt is made to return details in the body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
...
components:
  schemas:
...
    GroupList:
      type: object
      required:
        - groups
      properties:
        groups:
          type: array
          items:
            $ref: '#/components/schemas/Group'
    Group:
      type: object
      required:
        - id
        - name
        - owner
      properties:
        id:
          type: string
          description: The internal primary key of the group, in the form of a UUID. Only set when reading groups, completely ignored when you send a group to us.
          example: 604f5ea8-d146-4aac-9a15-4dc33a84eb59
        name:
          type: string
          description: The name of the group, must be unique, but otherwise just used for display purposes
          example: Kittens
          maxLength: 80
          minLength: 1
        flags:
          type: array
          items:
            type: string
          description: A list of flags as declared in configuration. Flags are used to store yes/no-style information about the group such as "wheelchair", etc.
          example:
            - privacy
            - wheelchair
        comments:
          type: string
          description: Optional comments the owner wishes to make regarding the group. Not processed in any way.
          example: A personal reason for the privacy request
        maximum_size:
          type: integer
          minimum: 0
          description: if set higher than 0 (the default), will limit the number of people that can join the group. Note that there is also a configuration item that globally limits the size of groups, e.g. to the maximum room size.
          example: 6
        owner:
          type: integer
          description: the badge number of the group owner. Must be a member of the group. If you are not an admin, you can only create groups with yourself as owner.
        members:
          type: array
          description: the current group members. READ ONLY, provided for ease of use of the API, but completely ignored in all write requests. Please use the relevant subresource API endpoints to manipulate group membership.
          items:
            $ref: '#/components/schemas/Member'
        invites:
          type: array
          description: the current outstanding invites for this group. READ ONLY, provided for ease of use of the API, but completely ignored in all write requests. Please use the relevant subresource API endpoints to send/revoke invites.
          items:
            $ref: '#/components/schemas/Member'
    GroupCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: The name of the group, must be unique, but otherwise just used for display purposes
          example: Kittens
          maxLength: 80
          minLength: 1
        flags:
          type: array
          items:
            type: string
          description: A list of flags as declared in configuration. Flags are used to store yes/no-style information about the group such as "wheelchair", etc.
          example:
            - public
            - wheelchair
        comments:
          type: string
          description: Optional comments the owner wishes to make regarding the group. Not processed in any way.
          example: A personal reason for the privacy request
...
    Member:
      type: object
      required:
        - id
        - nickname
        - hasKey
      properties:
        id:
          type: integer
          description: badge number (id in the attendee service).
          example: 42
        nickname:
          type: string
          description: The nickname of the attendee, proxied from that attendee service.
          example: Jumpy
        avatar:
          type: string
          description: A url to obtain the avatar for this attendee, points to an image such as a png or jpg. May require the same authentication this API expects.
          example: 'http://example.com/some/avatar.png'
        hasKey:
          type: boolean
          default: false
          description: Set to true if this person has been given a key to the room, for groups this can only be set if already assigned a room.
    Error:
      type: object
      required:
        - message
        - timestamp
        - requestid
      properties:
        timestamp:
          type: string
          format: date-time
          description: The time at which the error occurred.
          example: 2006-01-02T15:04:05+07:00
        requestid:
          type: string
          description: An internal trace id assigned to the error. Used to find logs associated with errors across our services. Display to the user as something to communicate to us with inquiries about the error.
          example: a8b7c6d5
        message:
          type: string
          description: |-
            A keyed description of the error. We do not write human readable text here because the user interface will be multi language.
            
            At this time, there are these values:
            - auth.forbidden (permissions missing or not a registered attendee)
            - auth.unauthorized (token missing completely or invalid)
            - group.data.duplicate (group with same name already exists, cannot create or rename)
            - group.data.invalid (invalid field contents)
            - group.id.invalid (invalid uuid id format)
            - group.id.notfound (no such group id)
            - group.invite.mismatch (nickname did not match - invite not sent)
            - group.mail.error (mail service reported error when sending notification)
            - group.member.duplicate (attendee is already in another group or has been assigned to a room already)
            - group.member.notfound (attendee is not in any group, or attendee not found when trying to add)
            - group.owner.notfound (requested owner is not part of this group)
            - group.owner.cannot.remove (this attendee is currently the owner of the group. Either change the owner first, or disband the group completely)
            - group.parse.error (invalid json)
            - group.read.error (database error)
            - group.write.error (database error)
            - room.data.duplicate (room with same name already exists, cannot create or rename)
            - room.data.invalid (invalid field contents)
            - room.group.duplicate (group is already assigned to another room)
            - room.group.notfound (group is not in this room, or group not found when trying to add/remove)
            - room.id.invalid (invalid uuid id format)
            - room.id.notfound (no such room id)
            - room.member.duplicate (attendee is already in another room)
            - room.member.in.group (attendee is in a group, cannot assign individually - assign the whole group or remove from group)
            - room.member.notfound (attendee is not in any/this room, or attendee not found when trying to add/remove)
            - room.not.empty (cannot delete a room that isn't empty)            
            - room.parse.error (invalid json)
            - room.read.error (database error)
            - room.too.small (not enough space in room to add individual or group)
            - room.write.error (database error)
          example: group.read.error
        details:
          type: object
          description: Optional additional details about the error. If available, will usually contain English language technobabble.
          additionalProperties:
            type: array
            items:
              type: string
          example:
            other:
              - no such group 'kerfluffle'
              - you need to refill the flux capacitor before the operation can succeed
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: A bearer or session token obtained from the auth service (which may integrate an OIDC identity provider)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: A shared secret used for local communication (also useful for local development)